+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add `purge` parameter for `text_progressive` (@xvw, reported by @Princesseuh)
 - Add command `abs` (@xvw, reported by @Zer0xxxx)
 - Add command `include_common_event` (@BastienDuplessier)
+- Add command `get_squares_between` and `get_squares_between_events` (@BastienDuplessier)
 
 ### Bug fixes
 - Fix Game_End background that wasn't disposed on `skip_title` activated (@BastienDuplessier, reported by @Zer0xxxx)